requirements.txt
Flask
flask-cors
python-dotenv
pymysql
Pillow
openai
requests